refactor login page, css, completed groceries page

+++ src/components/groceryCard/GroceryCard.js
@@ -1,15 +1,13 @@
 import gc from './groceryCard.module.css'
 import remove from '../../images/icons/delete.svg'
 import removeGrocery from '../../api/grocery';
-import { useAuth } from '../../providers/AuthProvider';
-import { fetchUser } from '../../api/user';
-export default function GroceryCard({data,onDelete}) {
-  
- const {getUserData,setUserData} = useAuth();
-  if (!data) return null;
- const leftItemsIcons = {};
- const rightItemsIcons = [{src : remove, alt : "Remove", clickaction : removeGroceryCall}]
 
+export default function GroceryCard({data,onDelete,onClick}) {
+
+  const leftItemsIcons = {};
+  const rightItemsIcons = [{src : remove, alt : "Remove", clickaction : removeGroceryCall}]
+  
+  if (!data) return null;
  function parseDate(date) {
     try {
     return date.toDate().toLocaleDateString('en-CA', {
@@ -24,8 +22,11 @@ export default function GroceryCard({data,onDelete}) {
 }
 
 async function removeGroceryCall() {
-  await removeGrocery(data.owner,data.id);
-   onDelete && onDelete();
+  let res = window.confirm("Delete Grocery ?")
+  if (res){
+    await removeGrocery(data.owner,data.id);
+    onDelete && onDelete();
+  }
 }
 
   return (
@@ -36,13 +37,13 @@ async function removeGroceryCall() {
            <img  key={index} src={item.src} className={gc.menuIcon} onClick={item.clickaction} alt={item.alt} />
              ))
          }
-        <div className={gc.dataWrapper}>
+         </div>
+        <div className={gc.dataWrapper} onClick={onClick}>
             <div className={`${gc.label} ${data.type === "shared" ? gc.colorShared : gc.colorPersonal}`}>{data.type} </div>
             <div className={gc.name}>{data.name}</div> 
             <div className={gc.date}>{parseDate(data.date)}</div> 
             <div className={`${gc.status}`}><span className={data.status==="completed" ? gc.colorCompleted : gc.colorPending}>{data.status}</span></div> 
        </div>
-        </div>
         <div className={gc.rightItems}>
          {rightItemsIcons && 
             rightItemsIcons.map((item, index) => (
